chore: fix e2e cli
