+++ .npmignore
@@ -13,8 +13,8 @@ src/
 .vscode/
 .circleci/
 debug/
-
-
+docker-examples/
+reports/
 ## assets and website
 assets/
