+++ .prettierignore
@@ -27,3 +27,5 @@ test/functional/store/*
 storage_default_storage/*
 docker-examples/
 .prettierignore
+.npmignore
+.gitignore
